refactor: make localhost API fallback a CLI-wide API base URL resolver
Replace the login-only getLoginApiBaseUrl with a shared getApiBaseUrl in
console-url.ts and route the ApifyClient creation in utils.ts through it, so the
"localhost Console => local API on :3333" default applies to every command, not
just login. An explicit APIFY_CLIENT_BASE_URL still overrides it.

As a result, setting APIFY_CONSOLE_URL to a localhost instance now also points
the API at localhost by default across the whole CLI. Docs and tests updated to
match; the login command no longer needs its own resolver.
